MINOR: mt_list: create a minimal initial include file
The goal is to have the minimal contents needed to accept the mt_list
patches in a way that builds. Here we're only defining a minimal generic
mt_list_cpu_relax() function and the MT_ALREADY_CHECKED() macro to hide a
variable's provenance. It also includes inttypes.h for portable stdint
definitions.