MINOR: implement some basic exponential back-off
Without EBO a 16-thread test can take 40s to complete and a 56-thread
one seems never to complete. With EBO it's much better. However, it is
observed that using too low a base or too low a limit to the loop count
is counter-productive on large-scale systems, while using too large of
any is counter-productive on laptop-type processors. The 8+7 base that
makes the loop count grow by 3 bits per failure seems like a good
balance in tests.