MINOR: rename mt_list_reconnect() to mt_list_connect_ends()
The difference in signification between connect_elem() and reconnect()
was not obvious, especially when seeking which function to use to
reconnect an element that was temporarily extracted. Here we're really
aiming at connecting both ends together to complete a removal, so let's
call the function mt_list_connect_ends(), which is better aligned with
the other one (mt_list_connect_elem()).