MINOR: clarify the validity domain of lock/unlock next/prev.
prev is slightly stricter than next so as to drop a check because it's
only used in a forward list_for_each(). We don't have any backwards
equivalent (both would deadlock) so that's fine. The function are not
strictly limited to iterators, though, which is why it's better to
mention clearly what is expected from the caller instead of reserving
them to iterators.

While we're at it, let's document these functions even if not intended
for use outside the loops, and clarify a few related ambiguities in the
doc.