MEDIUM: api: rename mt_list_cut_before() to mt_list_lock_prev()
The previous name wasn't very clear about what exactly was cut. Now by
introducing the notion of lock in it, we expect the operation to be
atomic and to be undoable using the unlock equivalent.

Note that we now we have two variants of these functions:
  - one that returns a complete element with both ends, one of
    which is the element itself
  - one that returns only the pointer to the next element before
    the link was locked.

In addition the second performs a read-only check for locked before
the CAS while the former does not. Maybe we could unify these ?