MEDIUM: api: rename mt_list_connect_elem() to mt_list_unlock_full()
The main use case of that function is to unlock an element that was
previously locked using mt_list_lock_full(), so better respect the
symmetry of the calls and call it mt_list_unlock_full() so as to make
the code easier to follow. If other use cases appear it will be easy
to make an alias for it.