MINOR: api: add mt_list_lock_elem(), mt_list_unlock{elem,self}()
These ones offer the ability to lock an element without locking its links.
This is normally not directly used, however the unlock operation will be
used because list iterators will definitely present locked elements that
the user might want to unlock. The _self() variant really is for use
inside iterators. Instead of restoring the original element (since its
values were both replaced with BUSY), it will loop the element over itself
and effectively unlock it. It's unclear at this point if there could be
other uses of this function which has the benefit of unlocking another
user of the same element which will also notice it is detached. Maybe a
better name will have to be found again in the future.