MINOR: introduce the MT_LIST_FOR_EACH_ENTRY_UNLOCKED() iterator
Now we have two explicit versions: locked and unlocked. The locked one
(historical one, with word "safe" dropped) continues to return a locked
element that the caller might have to explicitly unlock. This new one
pre-unlocks the element before the loop block, and restores lm.next to
BUSY after the loop block, so that the user doesn't have to care about
it. But it is only usable when the element cannot vanish.

This will permit to make a simpler flavor of the loop construct that
doesn't require to keep the element locked nor to unlock it when removing
it, and that instead will loop it over itself and assume that it is not
being used in parallel. There are definitely situations where this can
make sense (e.g. MPMC queues where elements cannot be removed by other
methods).

This way the locked variant will be explicit and it will feel more
natural to have to unlock it when deleting it, and the unlocked one
(even if strictly speaking less efficient) will be suitable for all
situations where one does not really care.