BUILD: refine MT_LIST_ELEM() so as to please gcc-13
Building with gcc-13.2 can yield some bogus warnings about access out of
bounds when assigning stuff to <item> in iterators despite the element
always perfectly defined where this happens. It seems it's confused by
the for() loop construct. It turns out that casting both sides of
MT_LIST_ELEM() to size_t is sufficient to make it happy. Better add a
last one so that the result is always unsigned (in case it would complain
on some platforms that the pointer type is not compatible). Unfortunately
this conversion from pointers to ints has a cost in that it adds a few
unneeded instructions, but gcc being what it is nowadays, there's not
much hope anymore to write efficient *and* safe code.