BUILD: support gcc before 4.7
Gcc before 4.7 did not have the __atomic_* API. The code currently uses
very little of it, in practice, only __atomic_exchange_n() needs to be
implemented (like for tcc), and that one can be done using a CAS that was
present in the older __sync_* API, so let's do that to ease portability
with older compilers (at least for regression testing purposes). This was
tested with gcc-4.4 on x86_64.