CLEANUP: better work around for the incorrect null-deref warning
Gcc used to believe about a potential null deref in the iterators because
it didn't know that there was necessarily an exclusion between item==NULL
and back.prev=NULL given that back.prev may only be null during the first
pass and never later. Let's place an MT_UNREACHABLE() on this condition
to let it know about it. It gets rid of the warning and slightly reduces
the code by letting the compiler aware of all possible conditions instead
of hiding the provenance of the pointers via a dummy asm() statement. We
just need to check for clang or a recent enough gcc so that we don't
break build on tcc and older compilers.